Merge initial state from new components with current state
When replacing a proxied component that has no state with
one that does have state, renders and lifecycle methods that
depend on the initial state will typically fail with
'this.state is null' errors.

This addresses the issue by invoking getInitialState() on
classic components or running the constructor for modern components
and merging the resulting state with the existing state, with
the existing state taking priority.

Fixes #38